Remove audio data test
Originally noted in #721, the current test only checks for audio
elements on the window. This results in true in pretty much all
browsers. Audio data only exists in FF as a depricated API. The test is
currently broken. I created a new pull since the old one (#721) was old
and out of date.