Detect emoji support on Windows 7+
U+1F603 SMILING FACE WITH OPEN MOUTH is just an outline on Windows, so we
weren't detecting that anything was drawn at all. Now we use U+1F428 KOALA
instead, which is filled in by the Windows emoji font (Segoe UI Symbol). Since
these emoji are monochrome on Windows 7/8, we set the fill color to red to
ensure we get a non-zero red value from the pixel we sample.