History API unavailable for URLs with 'file:' protocol
Browsers that usually support the History API will disable it when
the current URL is served off the filesystem with the 'file:'
protocol.

By including this protocol check, this becomes a more accurate test for
apps running in Cordova/PhoneGap/similar, as HTML files in these
scenarios are served locally off the device filesystem (so with
location.protocol === 'file:'). Although the browser used by the
Cordova app may support the history API, it doesn't with the 'file:'
protocol.