update invalid datachannel detect
the original detect passed in chrome, but was not technically valid. In
speaking with colleagues on Edge, this change makes a feature detect
that prevents possible errors from occurring

with '#' will be ignored, and an empty message aborts the commit. # On
branch master # Your branch is up-to-date with 'upstream/master'. # #
Changes to be committed: #	modified:
feature-detects/webrtc/datachannel.js #