Remove ancient (and expensive) logic for verifying CSS transform support.
This code was added in 2009 via [1] to work around a false-positive in WebKit.
Per [2] both Safari and Chrome have supported transforms since 2009 and 2011
respectively, and current global usage of earlier versions of those products
is negligible. Also note that this code originally attempted to limit itself
to WebKit, but the prefixed properties it detects have since been standardized
and implemented in all browsers.

This code is responsible for significant slowdowns (in the tens of tens of
milliseconds) when loading CNN.

[1] https://github.com/Modernizr/Modernizr/issues/15
[2] https://caniuse.com/#search=perspective