remove touchevents
This PR removes the `touchevents` detect from the project.

While renaming the test did, in fact, make the test accurately describe what it was testing, it still doesn't do what people think it does. this creates endless confusion (more than we see) and bug reports (the tip of the iceberg of people having problems with this detect.)

We can't create a test that does what people think this test does so we should end the confusion and remove the test. If people want the test in their projects it's an easy one to implement.

Additionally, I would argue that testing for what people _think_ this test does is harmful to users and contributing to that pattern (even via a misunderstanding of what this test does) is bad for the web.