indexeddbblob: Fix test never triggering
The `indexedbblob` test never triggers, even on platforms that support
IndexedDB or blobs in IndexedDB databases.

This is because in 164e757, `indexeddb` was changed to an async test,
but `indexeddbblob` still assumes that `indexeddb` test is synchronous.
Therefore, the `indexeddb` check is never true and `indexedbblob` never
triggers.

The `onsuccess` / `onerror` / exception handlers also assign the test
result to a shared variable before calling `addTest()`, which results
in the test result being always false as the variable will not be
updated synchronously before `addTest()`.

Let's:

* wrap the `indexeddbblob` test in a `on('indexeddb.deletedatabase')`
listener, so that it will trigger when `indexeddb` and
`indexeddb.deletedatabase` pass

* call `addTest()` in the `onsuccess` / `onerror` / exception handlers
instead of assigning to a shared variable