Update flexgap.js (#2749)
* Update flexgap.js

Added a check for the type of result: In the previous code, there was no check for the type of result. In the updated code, I added a check to ensure that result is a string, and if not, an error message is logged. 4

* Update flexgap.js

Added a check for Safari bug:

var isSupported = flexHeight === 1 || flexHeight === 2;
Reason: Safari, at zoom levels of 50% and 150%, returns scrollHeight as 2 instead of 1 for flex containers. This line ensures the test works across different browsers, including buggy Safari versions.

---------

Co-authored-by: veeck <gitkraken@veeck.de>